Moved config to utils
